+++ AquaMate/UI/Dialogs/SnapshotEditDlg.Designer.cs
@@ -11,6 +11,7 @@
         private System.Windows.Forms.Button btnLoad;
         private System.Windows.Forms.DateTimePicker dtpDateTime;
         private System.Windows.Forms.Label lblDate;
+        private System.Windows.Forms.Button btnSave;
         
         protected override void Dispose(bool disposing)
         {
@@ -32,15 +33,16 @@
             this.btnLoad = new System.Windows.Forms.Button();
             this.dtpDateTime = new System.Windows.Forms.DateTimePicker();
             this.lblDate = new System.Windows.Forms.Label();
+            this.btnSave = new System.Windows.Forms.Button();
             ((System.ComponentModel.ISupportInitialize)(this.pictureBox1)).BeginInit();
             this.SuspendLayout();
             // 
             // btnAccept
             // 
-            this.btnAccept.Location = new System.Drawing.Point(172, 351);
+            this.btnAccept.Location = new System.Drawing.Point(291, 343);
             this.btnAccept.Margin = new System.Windows.Forms.Padding(2);
             this.btnAccept.Name = "btnAccept";
-            this.btnAccept.Size = new System.Drawing.Size(120, 30);
+            this.btnAccept.Size = new System.Drawing.Size(96, 24);
             this.btnAccept.TabIndex = 6;
             this.btnAccept.Text = "Accept";
             this.btnAccept.TextImageRelation = System.Windows.Forms.TextImageRelation.ImageBeforeText;
@@ -50,10 +52,10 @@
             // btnCancel
             // 
             this.btnCancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.btnCancel.Location = new System.Drawing.Point(297, 351);
+            this.btnCancel.Location = new System.Drawing.Point(391, 343);
             this.btnCancel.Margin = new System.Windows.Forms.Padding(2);
             this.btnCancel.Name = "btnCancel";
-            this.btnCancel.Size = new System.Drawing.Size(120, 30);
+            this.btnCancel.Size = new System.Drawing.Size(96, 24);
             this.btnCancel.TabIndex = 7;
             this.btnCancel.Text = "Cancel";
             this.btnCancel.TextImageRelation = System.Windows.Forms.TextImageRelation.ImageBeforeText;
@@ -61,19 +63,19 @@
             // 
             // lblName
             // 
-            this.lblName.Location = new System.Drawing.Point(14, 18);
+            this.lblName.Location = new System.Drawing.Point(11, 14);
             this.lblName.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.lblName.Name = "lblName";
-            this.lblName.Size = new System.Drawing.Size(132, 21);
+            this.lblName.Size = new System.Drawing.Size(106, 17);
             this.lblName.TabIndex = 4;
             this.lblName.Text = "Name";
             // 
             // txtName
             // 
-            this.txtName.Location = new System.Drawing.Point(151, 14);
-            this.txtName.Margin = new System.Windows.Forms.Padding(2, 2, 2, 9);
+            this.txtName.Location = new System.Drawing.Point(121, 11);
+            this.txtName.Margin = new System.Windows.Forms.Padding(2, 2, 2, 7);
             this.txtName.Name = "txtName";
-            this.txtName.Size = new System.Drawing.Size(262, 26);
+            this.txtName.Size = new System.Drawing.Size(210, 22);
             this.txtName.TabIndex = 5;
             // 
             // pictureBox1
@@ -81,20 +83,20 @@
             this.pictureBox1.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.pictureBox1.ErrorImage = null;
             this.pictureBox1.InitialImage = null;
-            this.pictureBox1.Location = new System.Drawing.Point(14, 88);
-            this.pictureBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 9);
+            this.pictureBox1.Location = new System.Drawing.Point(11, 70);
+            this.pictureBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 7);
             this.pictureBox1.Name = "pictureBox1";
-            this.pictureBox1.Size = new System.Drawing.Size(400, 252);
+            this.pictureBox1.Size = new System.Drawing.Size(476, 264);
             this.pictureBox1.SizeMode = System.Windows.Forms.PictureBoxSizeMode.Zoom;
             this.pictureBox1.TabIndex = 8;
             this.pictureBox1.TabStop = false;
             // 
             // btnLoad
             // 
-            this.btnLoad.Location = new System.Drawing.Point(11, 351);
+            this.btnLoad.Location = new System.Drawing.Point(11, 343);
             this.btnLoad.Margin = new System.Windows.Forms.Padding(2);
             this.btnLoad.Name = "btnLoad";
-            this.btnLoad.Size = new System.Drawing.Size(120, 30);
+            this.btnLoad.Size = new System.Drawing.Size(96, 24);
             this.btnLoad.TabIndex = 6;
             this.btnLoad.Text = "Load";
             this.btnLoad.TextImageRelation = System.Windows.Forms.TextImageRelation.ImageBeforeText;
@@ -105,34 +107,47 @@
             // 
             this.dtpDateTime.CustomFormat = "dd.MM.yyyy HH:mm:ss";
             this.dtpDateTime.Format = System.Windows.Forms.DateTimePickerFormat.Custom;
-            this.dtpDateTime.Location = new System.Drawing.Point(151, 51);
-            this.dtpDateTime.Margin = new System.Windows.Forms.Padding(2, 2, 2, 9);
+            this.dtpDateTime.Location = new System.Drawing.Point(121, 41);
+            this.dtpDateTime.Margin = new System.Windows.Forms.Padding(2, 2, 2, 7);
             this.dtpDateTime.Name = "dtpDateTime";
-            this.dtpDateTime.Size = new System.Drawing.Size(177, 26);
+            this.dtpDateTime.Size = new System.Drawing.Size(142, 22);
             this.dtpDateTime.TabIndex = 10;
             // 
             // lblDate
             // 
-            this.lblDate.Location = new System.Drawing.Point(14, 57);
+            this.lblDate.Location = new System.Drawing.Point(11, 46);
             this.lblDate.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.lblDate.Name = "lblDate";
-            this.lblDate.Size = new System.Drawing.Size(132, 21);
+            this.lblDate.Size = new System.Drawing.Size(106, 17);
             this.lblDate.TabIndex = 9;
             this.lblDate.Text = "Date";
+            // 
+            // btnSave
+            // 
+            this.btnSave.Location = new System.Drawing.Point(111, 343);
+            this.btnSave.Margin = new System.Windows.Forms.Padding(2);
+            this.btnSave.Name = "btnSave";
+            this.btnSave.Size = new System.Drawing.Size(96, 24);
+            this.btnSave.TabIndex = 6;
+            this.btnSave.Text = "Save";
+            this.btnSave.TextImageRelation = System.Windows.Forms.TextImageRelation.ImageBeforeText;
+            this.btnSave.UseVisualStyleBackColor = true;
+            this.btnSave.Click += new System.EventHandler(this.btnSave_Click);
             // 
             // SnapshotEditDlg
             // 
             this.AcceptButton = this.btnAccept;
-            this.AutoScaleDimensions = new System.Drawing.SizeF(120F, 120F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(96F, 96F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
             this.CancelButton = this.btnCancel;
-            this.ClientSize = new System.Drawing.Size(428, 392);
+            this.ClientSize = new System.Drawing.Size(498, 378);
             this.Controls.Add(this.dtpDateTime);
             this.Controls.Add(this.lblDate);
             this.Controls.Add(this.pictureBox1);
             this.Controls.Add(this.txtName);
             this.Controls.Add(this.lblName);
             this.Controls.Add(this.btnCancel);
+            this.Controls.Add(this.btnSave);
             this.Controls.Add(this.btnLoad);
             this.Controls.Add(this.btnAccept);
             this.Font = new System.Drawing.Font("Segoe UI", 8.25F);